+++ tests/test_new_qiskit_temp.py
@@ -2,33 +2,39 @@ import pytest
 import pennylane as qml
 import qiskit
 
+from semantic_version import Version
 from unittest.mock import Mock
 
+from pennylane_qiskit import BasicSimulatorDevice
 
-@pytest.mark.parametrize(
-    "device_name",
-    [
-        "qiskit.aer",
-        "qiskit.basicaer",
-        "qiskit.remote",
-    ],
+
+@pytest.mark.skipif(
+    Version(qiskit.__version__) < Version("1.0.0"),
+    reason="versions below 1.0 are compatible with BasicAer",
 )
-def test_error_is_raised_if_initalizing_device(monkeypatch, device_name):
+def test_error_is_raised_if_initalizing_basicaer_device(monkeypatch):
     """Test that when Qiskit 1.0 is installed, an error is raised if you try
-    to initialize a device. This is a temporary test and will be removed along
-    with the error one everything is compatible with 1.0"""
-
-    # make it not fail in the normal tests (running 0.46)
-    if qiskit.__version__ != "1.0.0":
-        monkeypatch.setattr(qiskit, "__version__", "1.0.0")
+    to initialize the 'qiskit.basicaer' device."""
 
     # test that the correct error is actually raised in Qiskit 1.0 (rather than fx an import error)
     with pytest.raises(
         RuntimeError,
-        match="The devices in the PennyLane Qiskit plugin are currently only compatible with versions of Qiskit below 0.46",
+        match="Qiskit has discontinued the BasicAer device",
     ):
-        if device_name in ["qiskit.aer", "qiskit.basicaer"]:
-            qml.device(device_name, wires=2)
-        else:
-            # use a Mock backend to avoid call to the remote service
-            qml.device(device_name, wires=2, backend=Mock())
+        qml.device("qiskit.basicaer", wires=2)
+
+
+@pytest.mark.skipif(
+    Version(qiskit.__version__) >= Version("1.0.0"),
+    reason="versions 1.0 and above are compatible with BasicSimulator",
+)
+def test_error_is_raised_if_initalizing_basic_simulator_device(monkeypatch):
+    """Test that when a version of Qiskit below 1.0 is installed, an error is raised if you try
+    to initialize the BasicSimulatorDevice."""
+
+    # test that the correct error is actually raised in Qiskit 1.0 (rather than fx an import error)
+    with pytest.raises(
+        RuntimeError,
+        match="device is not compatible with version of Qiskit prior to 1.0",
+    ):
+        BasicSimulatorDevice(wires=2)