+++ CHANGELOG.md
@@ -2,7 +2,21 @@
 
 ### New features since last release
 
+* Support is added for using the plugin devices with Qiskit 1.0. As the backend provider ``qiskit.BasicAer`` 
+  is no longer supported by Qiskit in 1.0, this added support does not extend to the ``"qiskit.aer"`` device. 
+  Instead, a ``"qiskit.basicsim"`` device is added, with the new Qiskit implementation of a Python simulator 
+  device, ``BasicSimulator``, as the backend.
+  [(#493)](https://github.com/PennyLaneAI/pennylane-qiskit/pull/493)
+
 ### Improvements 🛠
+
+* Following updates to allow device compatibility with Qiskit 1.0, the version of `qiskit-ibm-runtime` is 
+  no longer capped.
+  [(#508)](https://github.com/PennyLaneAI/pennylane-qiskit/pull/508)
+
+* The test suite now runs with the most recent `qiskit` and `qiskit-ibm-runtime`, and well as with 
+  `'qiskit==0.45'` and `qiskit-ibm-runtime<0.21` to monitor backward-compatibility.
+  [(#508)](https://github.com/PennyLaneAI/pennylane-qiskit/pull/508)
 
 ### Breaking changes 💔
 
@@ -15,6 +29,7 @@
 ### Contributors ✍️
 
 This release contains contributions from (in alphabetical order):
+Lillian M. A. Frederiksen
 
 ---
 # Release 0.35.1